requirements.txt
easydict==1.10
matplotlib==3.7.1
numpy==1.24.2
scikit_learn==1.2.2
scipy==1.10.1
Shapely==2.0.1
skimage==0.0
tqdm==4.65.0
hydra-core==1.3.2
